Refactor codebase to align with updated Go version support
- Removed Go 1.20 specific files and tests to clean up legacy code.
- Consolidated tests for newer Go versions to streamline maintainability.
- Updated logging and debug-related test cases to improve clarity and consistency.
- Migrated `SendWithSMTPClient` method implementation and documentation to a new consolidated structure.
