+++ smtp/smtp_test.go
@@ -3302,6 +3302,39 @@ func TestClient_SetDebugLog(t *testing.T) {
 	})
 }
 
+func TestClient_SetDebugLog_JSON(t *testing.T) {
+	t.Run("set debug loggging to on should not override logger", func(t *testing.T) {
+		client := &Client{logger: log.NewJSON(os.Stderr, log.LevelDebug)}
+		client.SetDebugLog(true)
+		if !client.debug {
+			t.Fatalf("expected debug log to be true")
+		}
+		if client.logger == nil {
+			t.Fatalf("expected logger to be defined")
+		}
+		if !strings.EqualFold(fmt.Sprintf("%T", client.logger), "*log.JSONlog") {
+			t.Errorf("expected logger to be of type *log.JSONlog, got: %T", client.logger)
+		}
+	})
+}
+
+func TestClient_SetLogger_JSON(t *testing.T) {
+	t.Run("set logger to JSONlog logger", func(t *testing.T) {
+		client := &Client{}
+		client.SetLogger(log.NewJSON(os.Stderr, log.LevelDebug))
+		if !strings.EqualFold(fmt.Sprintf("%T", client.logger), "*log.JSONlog") {
+			t.Errorf("expected logger to be of type *log.JSONlog, got: %T", client.logger)
+		}
+	})
+	t.Run("nil logger should just return and not set/override", func(t *testing.T) {
+		client := &Client{logger: log.NewJSON(os.Stderr, log.LevelDebug)}
+		client.SetLogger(nil)
+		if !strings.EqualFold(fmt.Sprintf("%T", client.logger), "*log.JSONlog") {
+			t.Errorf("expected logger to be of type *log.JSONlog, got: %T", client.logger)
+		}
+	})
+}
+
 func TestClient_SetLogger(t *testing.T) {
 	t.Run("set logger to Stdlog logger", func(t *testing.T) {
 		client := &Client{}
